requirements.txt
python-dotenv==1.1.1
fastapi==0.120.0
uvicorn==0.38.0
pydantic==2.12.3
requests==2.32.5
firebase-admin==7.1.0
google-cloud-firestore==2.21.0
google-genai==1.50.1
ipykernel==7.1.0
beautifulsoup4==4.14.2
faiss-cpu==1.12.0
langchain-openai==1.0.3
langchain-community==0.4.1